CRUD care diary feature

+++ src/main/java/com/bach/qlkh/repository/ManagerRepository.java
@@ -3,5 +3,5 @@ package com.bach.qlkh.repository;
 import com.bach.qlkh.model.Manager;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ManagerRepository extends JpaRepository<Manager,Integer> {
+public interface ManagerRepository extends JpaRepository<Manager,Long> {
 }
